Moved AuthorWrapper and PubMedArticleWrapper into their own files.
